backend/requirements.txt
Django==3.1.3
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
django-cors-headers==3.10.0
